+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fix parsing of `<ruby>` tags, especially `<rb>` and `<rtc>`.
 * Create `HTMLMenuItemElement` when parsing `<menuitem>` tags.
 * Create `HTMLSourceElement` when parsing `<source>` tags.
+* Create `HTMLTrackElement` when parsing `<track>` tags.
 
 # domino 1.0.23 (30 Jan 2016)
 * Fix `CSSStyleDeclaration#setProperty`. (#71)